core: heartbeat: Expire known peers after consecutive failed heartbeats
This commit adds peer expiration to the gossip protocol. If a peer
has not sent a successful heartbeat within some tolerated time, we
expire the peer and no longer send it heartbeats.

This will eventually result in a replication event sent to other
cluster members to re-replicate the wallets owned by the failed peer.