circuits: benches: Add benchmark for small match circuit
The smaller circuit only checks the validity of a match; i.e. does
not compute the match in full. It also checks that the match results
from a user defined order and that the counterparties have the
balances to cover the match.

Adding a benchmark so we can track this circuit's proving performance
against the standard circuit.