circuits: mpc-gadgets: Improve numerical stability near field order
Previously, many of the gadgets had issues with numerical stability
for large values, particularly those over 2^252; which represent
small negative numbers (e.g. -1).

The general reason for this is that the Ristretto scalar field is
of prime order, not binary order; so for example a negative value
can have either the 251st bit or the 252nd bit set.