circuits: Fix integration tests by including crate patches
To override Arkworks crate versions to the version on remote master,
we add a `patches.crates-io` section to the workspace root cargo
config.

While adding patches to non-root workspace crates typically has no
effect; we run integration tests outside of the context of a
workspace. Therefore, we must configure the crate-local Cargo.toml
to run as a standalone crate; patches included.