core: state: Update global state to lock at finer granularity
The benefit of this approach is we get far less contention, especially
for cases in which we very infrequently need a write lock.

The downside is that we leave it up to the callsites to manage locks.
This has the potential to create race conditions if callsites do
not acquire locks in the same order as one-another. Long term we
should probably move the state locking into a dedicated lock manager
that orders the locks.