circuit-macros: Fix two small concurrency bugs
Technically, the tracer should always be run is a single-process,
serial manner; but it's easier to make it safe in the presence of
external concurrency guards.

The first bug comes in the implementation of the tracer epilogue;
the tracer should not clone the locked scope, it should directly
use the shared reference returned by the mutex to `scope_out`.

The second bug comes in the test suite, running tests concurrently
gives them the opportunity to layer their scopes on top of one
another. To fix this I added a shared `Mutex` to the start of each
test where the tracer is enabled.