circuits: types: keychain: Add specific allocated type for wallet keys
Many of the constraints involving the keys (e.g. ensuring that keys
haven't changed or authenticating proofs) involved indexing into an
array of keys. This is both not very clean and a pretty serious area
for bugs to creep in. Naming the keys directly makes these bugs harder
to write.