core: price-reporter: Pass API secrets via config
Previously we stored secrets in a .env file assumed to be placed
in the working directory of the relayer at startup. This change
unifies the config pattern used elsewhere in the relayer with the
config pattern used in the price reporter by using the manager
config to share API secrets.