circuits: zk-circuits: valid-match-mpc: Tune `BP_GENS_CAPACITY` constant
This contant should equal the number of constraints rounded up to the next
power of 2. These bulletproof generators take a non-trivial amount of time
to generate at runtime, so tuning them shaves off about 1.5s from the runtime
of the circuit.

After optimizing the circuit (removing input consistency checks) the resulting
constraint set is much smaller. So `BP_GENS_CAPACITY` can be reduced.

Down the line we can generate the generator chain ahead of time if we so prefer.