core: state: Simplify locking interface for callers
In specific, our state is slightly de-normalized and will likely
continue in that direction. This is largely for efficiency and
ease of data access.

Because the data is de-normalized, we'd like to provide a very
specific setter interface through which we can centrallize concurrency
management and ensure de-normalized data stays in sync.