circuits: types: Add proof linkability types
We link proofs together by guaranteeing input consistency across proofs.
This is done by committing to a witness variable using the same
randomness in each proof. The verifier can then use a single copy of
the input commitment that is supposed to be consistent.

This commit replaces many of the circuit types with `LinkableType`
equivalents, which track the randomness used in their commitment.