core: tasks: settle-match: Do not fail if counterparty encumbers first
When both relayers in a match attempt to submit a `match` transaction
to the contract, one will be sequenced first -- spending both nullifiers
and making the second transaction fail.

The relayer that submitted the failed `match` transaction may simply
abandon the `match` transaction as the counterparty has already
encumbered their wallet -- it is safe for both parties to settle.