core: api-server: router: Check signatures on raw request bytes
Deserializing and then re-serializing before we check signatures
modifies the request to (for example) remove whitespace, remove
leading zeros on biguints, etc. This makes it difficult for clients
to match serde's serializing to sign a consistent message. Instead
we can check the signature directly on the bytes from the req.