core: gossip: orderbook: Add order in `Received` state if invalid proof
Previously, we waited to validate the proof before indexing the order
in the book. The issue with this is that the heartbeat sync then causes
a continuous cycle of re-request and re-validate loops that end in
the same result.

Instead, we wait for the peer that published an invalid proof to send
a new, valid proof.