workers: network-manager: Multithread task handlers
We spawn network manager tasks on a tokio runtime instead of using a
single thread. This is complicated by `Swarm` not implementing `Sync`,
so we broker access to the `Swarm` via a `behavior_rx` queue. This queue
is drained and its events handled on the main thread to obviate
synchronization on the `Swarm`.