workers: task-driver: check if task is running before executing step logic (#1129)
In this PR, we check whether a task is running before retrying a task step. This prevents serial tasks on the same queue from executing simultaneously across a cluster in the case that one task was cleared from the queue due to e.g. a refresh while it was retrying a task step.

### Testing
- [x] All unit tests pass