darkpool-client: Use `quote_amount` from call in share parsing
If the contracts or a client decide to use a different quote
amount--which may be valid so long as its implied price is less than
that of the relayer authorized price--we recover the wrong shares. This
uses the exact quote amount swapped in the transaction to ensure we
recover the correct shares.